refactor(Money): 금액을 long으로 다룬다

+++ src/main/java/lotto/domain/Money.java
@@ -4,24 +4,24 @@ public class Money {
 
     private static final String NEGATIVE_AMOUNT_EXCEPTION_MESSAGE = "금액은 음수이면 안됩니다";
 
-    private final int amount;
+    private final long amount;
 
-    public Money(int amount) {
+    public Money(long amount) {
         validateGreaterThanZero(amount);
         this.amount = amount;
     }
 
-    private void validateGreaterThanZero(int amount) {
+    private void validateGreaterThanZero(long amount) {
         if (amount < 0) {
             throw new IllegalArgumentException(NEGATIVE_AMOUNT_EXCEPTION_MESSAGE);
         }
     }
 
-    public boolean hasChangesWith(int price) {
+    public boolean hasChangesWith(long price) {
         return amount % price != 0;
     }
 
-    public int countAffordableFor(int price) {
+    public long countAffordableFor(long price) {
         return amount / price;
     }
 }
